requirements.txt
feedparser~=6.0.11
boto3~=1.35.97
tweepy~=4.14.0
botocore~=1.35.97
DateTime~=5.5
facebook-sdk~=3.1.0
requests~=2.32.3
oauthlib~=3.2.2
instagram-private-api~=1.6.0
Pillow~=10.2.0